// src/components/molecules/Jitsi/types.tsx
export interface JitsiMeetAPI {
  // TODO, add definitions
}

export interface JitsiMeetAPIOptions {
  roomName?: string;

  width?: number;

  height?: number;

  parentNode?: ReactNode;

  configOverwrite?: ConfigOptions;

  interfaceConfigOverwrite?: InterfaceConfigOptions;

  noSSL?: boolean;

  jwt?: string;

  onLoad?: () => void;

  invites?: {};

  devices?: {
    audioInput?: string;
    audioOutput?: string;
    videoInput?: string;
  };

  userInfo?: {
    email?: string;
  };
}
// ...
export interface Props {
  /**
   * Main container styles
   */
  containerStyle?: React.CSSProperties;

  /**
   * Jitsi iframe container styles
   */
  loadingStyle?: React.CSSProperties;

  /**
   * Jitsi iframe container styles
   */
  frameStyle?: React.CSSProperties;

  /**
   * Component shown until JitsiMeet video conference is not started
   */
  loadingComponent?: React.ComponentType;

  /**
   * Callback function invoked with JitsiMeetExternalAPI object when it is loaded
   */
  onAPILoad?: (api: JitsiMeetAPI) => void;

  /**
   * Callback function invoked when JitsiMeet iframe is loaded
   */
  onIframeLoad?: () => void;

  /**
   * Domain used to build the conference URL, 'meet.jit.si' for example.
   */
  domain?: string;

  /**
   * Name of the room to join
   */
  roomName?: string;

  /**
   * Password to set for the meeting room
   */
  password?: string;

  /**
   * Participant's name
   */
  displayName?: string;

  // Options passed to JitsiMeetExternalAPI

  /**
   * Overrides for the default meeting options
   */
  config?: ConfigOptions;

  /**
   * Overrides for default meeting interface options
   */
  interfaceConfig?: InterfaceConfigOptions;

  /**
   * Boolean indicating if the server should be contacted using HTTP or HTTPS
   */
  noSSL?: boolean;

  /**
   * JWT token to pass to the domain
   */
  jwt?: string;

  /**
   * A map containing information about the initial devices that will be used in the call.
   */
  devices?: {
    audioInput?: string;
    audioOutput?: string;
    videoInput?: string;
  };

  /**
   * Object containing information about the participant opening the meeting, such as email
   */
  userInfo?: {
    email?: string;
    displayName?: string;
  };
}